PCIE -> Kernel Buffer Working
Unfortunately Kernel Buffer -> User Buffer is not.

I didn't understand that I can't copy data from a kernel buffer to a user
buffer from within an arbitrary kthread. This means that I can't use
workqueues to copy the data. Instead I need to think of a new mechanism
to send data to the user.

I think I figured it out.

I'll use a combination of a 2 count semaphore and a kernel fifo (kfifo)

When the read function is called it will initiate the read command,
and take two semaphors (1 for each buffer) then attempt to take
the semaphore. This will block as the user read thread is waiting
for the semaphores. The interrupt will populate a structure that is
enqueued into a kfifo then give a semaphore.

At this time the read thread will wake up, copy the data to the user
buffer take a semaphore and send the buffer to the driver to be used
by the PCIE device. This will continue until the user data is finished"